Extract generic RenderableViewZone hosting element
Move the view-zone hosting renderable out of embedded_comment.rs into a
shared app/src/editor/view_zone.rs module so non-comment zones can reuse
it. Adds a configurable x-origin mode: comments keep viewport-pinned
painting; content-space painting is available for full-width zones.
