+++ app/src/code/editor/embedded_comment.rs
@@ -8,7 +8,7 @@ use serde_yaml::Mapping;
 use uuid::Uuid;
 use warp_editor::content::markdown::MarkdownStyle;
 use warp_editor::editor::EmbeddedItemModel;
-use warp_editor::render::element::{RenderContext, RenderableBlock};
+use warp_editor::render::element::RenderableBlock;
 use warp_editor::render::layout::TextLayout;
 use warp_editor::render::model::viewport::ViewportItem;
 use warp_editor::render::model::{
@@ -16,16 +16,13 @@ use warp_editor::render::model::{
     LaidOutEmbeddedItem, RenderState,
 };
 use warpui::elements::ChildView;
-use warpui::event::DispatchedEvent;
 use warpui::units::Pixels;
-use warpui::{
-    AppContext, Element, EntityId, EventContext, LayoutContext, SizeConstraint, ViewHandle,
-    WindowId,
-};
+use warpui::{AppContext, Element, EntityId, ViewHandle, WindowId};
 
 use crate::code::editor::comment_editor::{CommentEditor, MAX_COMMENT_HEIGHT};
 use crate::code::editor::inline_comment_view::InlineCommentView;
 use crate::code_review::comments::CommentId;
+use crate::editor::view_zone::{RenderableViewZone, RenderableViewZoneSpacer, ViewZoneXOrigin};
 
 /// Upper bound on a saved inline comment card's laid-out height. Unlike the composer (capped at
 /// [`MAX_COMMENT_HEIGHT`] so its draft scrolls internally), a saved card reserves its full height so
@@ -39,13 +36,6 @@ const INLINE_COMMENT_FIRST_LINE_HEIGHT: f32 = 24.0;
 const COMMENT_ID_MAPPING_KEY: &str = "comment_id";
 const ENTITY_ID_MAPPING_KEY: &str = "entity_id";
 const WINDOW_ID_MAPPING_KEY: &str = "window_id";
-
-fn viewport_pinned_origin(viewport_item: &ViewportItem, ctx: &RenderContext) -> Vector2F {
-    let content_rect = viewport_item.content_bounds(ctx);
-    let mut origin = content_rect.origin();
-    origin.set_x(ctx.bounds.origin_x());
-    origin
-}
 
 #[derive(Debug)]
 pub struct EmbeddedCommentSpace {
@@ -189,10 +179,11 @@ impl LaidOutEmbeddedItem for LaidOutEmbeddedCommentSpace {
                 let child = ChildView::new(&editor).finish();
                 let window_id = self.window_id;
                 let entity_id = self.editor_entity_id;
-                Box::new(RenderableHostedComment::new(
+                Box::new(RenderableViewZone::new(
                     viewport_item,
                     child,
                     MAX_COMMENT_HEIGHT,
+                    ViewZoneXOrigin::ViewportPinned,
                     Box::new(move |measured, app| {
                         if let Some(editor) =
                             app.view_with_id::<CommentEditor>(window_id, entity_id)
@@ -202,7 +193,7 @@ impl LaidOutEmbeddedItem for LaidOutEmbeddedCommentSpace {
                     }),
                 ))
             }
-            None => Box::new(RenderableEmbeddedCommentSpace::new(viewport_item)),
+            None => Box::new(RenderableViewZoneSpacer::new(viewport_item)),
         }
     }
 
@@ -212,117 +203,6 @@ impl LaidOutEmbeddedItem for LaidOutEmbeddedCommentSpace {
 
     fn as_any(&self) -> &dyn Any {
         self
-    }
-}
-
-type SizeWriteBack = Box<dyn Fn(Vector2F, &AppContext)>;
-
-/// Hosts an inline comment view's element (either the active composer or a saved-comment card)
-/// inside a per-view inline comment block. It lays out the child against the block's content width
-/// (capped at `max_height`), writes the measured size back via `write_back_size` so the next layout
-/// reserves exactly the child's height, and paints and routes events to the child in content space
-/// (so it scrolls with its anchored line).
-pub struct RenderableHostedComment {
-    viewport_item: ViewportItem,
-    child: Box<dyn Element>,
-    max_height: f32,
-    write_back_size: SizeWriteBack,
-}
-
-impl RenderableHostedComment {
-    fn new(
-        viewport_item: ViewportItem,
-        child: Box<dyn Element>,
-        max_height: f32,
-        write_back_size: SizeWriteBack,
-    ) -> Self {
-        Self {
-            viewport_item,
-            child,
-            max_height,
-            write_back_size,
-        }
-    }
-}
-
-impl RenderableBlock for RenderableHostedComment {
-    fn viewport_item(&self) -> &ViewportItem {
-        &self.viewport_item
-    }
-
-    fn layout(&mut self, _model: &RenderState, ctx: &mut LayoutContext, app: &AppContext) {
-        let width = self.viewport_item.content_size.x();
-        let measured = self.child.layout(
-            SizeConstraint::new(vec2f(0., 0.), vec2f(width, self.max_height)),
-            ctx,
-            app,
-        );
-
-        (self.write_back_size)(measured, app);
-    }
-
-    fn paint(&mut self, _model: &RenderState, ctx: &mut RenderContext, app: &AppContext) {
-        let origin = viewport_pinned_origin(&self.viewport_item, ctx);
-        ctx.paint.scene.start_layer(warpui::ClipBounds::ActiveLayer);
-        self.child.paint(origin, ctx.paint, app);
-        ctx.paint.scene.stop_layer();
-    }
-
-    fn after_layout(&mut self, ctx: &mut warpui::AfterLayoutContext, app: &AppContext) {
-        self.child.after_layout(ctx, app);
-    }
-
-    fn dispatch_event(
-        &mut self,
-        _model: &RenderState,
-        event: &DispatchedEvent,
-        ctx: &mut EventContext,
-        app: &AppContext,
-    ) -> bool {
-        self.child.dispatch_event(event, ctx, app)
-    }
-
-    fn is_view_zone(&self) -> bool {
-        true
-    }
-}
-
-pub struct RenderableEmbeddedCommentSpace {
-    viewport_item: ViewportItem,
-}
-
-impl RenderableEmbeddedCommentSpace {
-    pub(crate) fn new(viewport_item: ViewportItem) -> Self {
-        Self { viewport_item }
-    }
-}
-
-impl RenderableBlock for RenderableEmbeddedCommentSpace {
-    fn viewport_item(&self) -> &ViewportItem {
-        &self.viewport_item
-    }
-
-    fn layout(&mut self, _model: &RenderState, _ctx: &mut LayoutContext, _app: &AppContext) {
-        // No-op: this is just a spacer, the actual editor is laid out by EditorWrapper
-    }
-
-    fn paint(&mut self, _model: &RenderState, _ctx: &mut RenderContext, _app: &AppContext) {
-        // No-op: this is just empty space, the actual editor is painted by EditorWrapper
-    }
-
-    fn dispatch_event(
-        &mut self,
-        _model: &RenderState,
-        _event: &DispatchedEvent,
-        _ctx: &mut EventContext,
-        _app: &AppContext,
-    ) -> bool {
-        // No interactivity: events are handled by the editor rendered by EditorWrapper
-        false
-    }
-
-    fn is_view_zone(&self) -> bool {
-        true
     }
 }
 
@@ -394,10 +274,11 @@ impl LaidOutEmbeddedItem for LaidOutInlineSavedComment {
                 let child = ChildView::new(&view).finish();
                 let window_id = self.window_id;
                 let entity_id = self.view_entity_id;
-                Box::new(RenderableHostedComment::new(
+                Box::new(RenderableViewZone::new(
                     viewport_item,
                     child,
                     UNBOUNDED_COMMENT_HEIGHT,
+                    ViewZoneXOrigin::ViewportPinned,
                     Box::new(move |measured, app| {
                         if let Some(view) =
                             app.view_with_id::<InlineCommentView>(window_id, entity_id)
@@ -407,7 +288,7 @@ impl LaidOutEmbeddedItem for LaidOutInlineSavedComment {
                     }),
                 ))
             }
-            None => Box::new(RenderableEmbeddedCommentSpace::new(viewport_item)),
+            None => Box::new(RenderableViewZoneSpacer::new(viewport_item)),
         }
     }
